add EuclideanKineticEnergy type
Different metrics require different integrators. We can warn users that
they are using the wrong integrator by creating a new
`EuclideanKineticEnergy` type. Euclidean metrics return a function that
computes the kinetic energy with this type; in exchange integrators for
euclidean metrics take this type as an input. This way mypy will warn
users when they try to use an integrator for euclidean metric with a
non-euclidean kinetic energy and conversely.

To ensure future compatibility with Riemannian metrics we pass the
position as an argument to "kinetic_energy" but ignore it for euclidean
metrics.