reorganize the hmc implementation
This commmit does not introduce anything new but re-organizes the
implementation of hmc to make it more general, and in particular allow
it to accomodate for the NUTS algorithms.

HMC in its more generic form can be understood as sampling a trajectory
and then sampling a proposal from this trajectory. In vanilla HMC we use
a deterministic expansion of the trajectory and accept or reject the
last state. The NUTS algorithm expands the trajectory randomly and
geometrically, updating the proposal at each step.