move hmc proposal to hmc algorithm file
The HMC proposal is currently implemented in a separate `algorithms.py`
file. This uselessly complicates the file structure of the library and
complicates the exploration. We thus move all the proposal logic and
related NamedTuples (HMCInfo) to the file that corresponds to the HMC
kernel. As a result, all there is to know about the HMC algorithm is
contained in one file.

We also move the function to create a new HMC state to the file
containing the definition.