Better residual resampling (#120)
This implementation of the residual resampler reuses our previous implementation of multinomial instead of the high memory footprint implementation of categorical sampling. The difference now lies in the use of a permutation of the multinomial indices to make sure that the sorted uniforms don't mess with the exchangeability properties. We increased the number of samples in the test as the algorithm now scales.