Allow users to pass a float for constant learning rate
The current API requires user to pass a function of the index of the
current iteration, even when the learning rate is constant. However, in
practice most people use a constant learning rate (even if they
shouldn't, but who are we to judge?)  and we should simplify that
pattern of use.

This commit thus allows users to pass either a function, either a
float as learning rate to the kernel builder. It will raise an
exception for any other type.