Build documentation in PRs
The documentation is currently only built when we merge on `main`, which
means that we need to trust that contributors (including ourselves) are
running the examples locally before pushing. This is too much to
ask (and cumbersome), so in this PR we edit the workflows so that docs
are built for each PR. They are only deployed when merging on `main`.