Remove support for custom `logprob_grad_fn`
Custom gradient currently need to be passed to the algorithms all the
way down to the integrators. However JAX has a much nicer way of doing
that; you can write a logprob and register a gradient for that logprob;
JAX will then use it when `jax.grad` is called on that function.

This simplifies the code in MEADS greatly: we not pass the logprob
gradients computed from GHMC and don't need to initialize the adaptation
with either the function that compute the gradient of the logprob, or a
batch function.