Pass the number of samples for resampling
The number of particles we resample in the SMC step is currently equal
to the number of weights passed to the resampling function. In this PR
we allow the caller to ask for a different number of particles. This
allows to build Waste-Free SMC kernels by asking to resample M < N
particles and build the update function so that it returns N particles.