Refactor adaptation to return parameters and extra info
We currently only return the last state, the values of the parameter and
the adapted kernel. However, the full chain and intermediate adaptation
states can be useful when debugging inference.

In addition, adaptation currently returns a `kernel` where the
parameters have already been specified. This is however a bit to high
level for Blackjax and can make vmap-ing adaptation difficult.

Finally, MEADS is currently only implemented as an adaptation scheme for
GHMC, we change its name to reflect this.

In this PR we make `window_adaptation`, `meads_adaptation` and
`pathfinder_adaptation` return extra information, and do not return the
kernel directly anymore.