Update typing in the library following current best practice (https://jax.readthedocs.io/en/latest/jax.typing.html).
General guideline:
- `ArrayLike` and `ArrayLikeTree` to annotate function input,
- `Array` and `ArrayTree` to annotate function output.

Leaves of a Pytree definition in the library are in principle annotated as `Array`, as they are mostly internal representation. For example:
```
class WelfordAlgorithmState(NamedTuple):
    mean: Array
    ...
```

Something I am not happy about is the way `logdensity` is annotated. While they are `Array` (as they are in most cases should be output of a Jax function), we annotate them as `float` to empathizes they should be scalar. I am leaving them as `float` for now until we introduce shape annotation.